test: 좋아요 기능 테스트
